ci: simply build all ports
after days of trial and error, it is not possible to build ports on
GitHub Action runners yet. The reason is the limited size of disk on
runners. The FreeBSD VM needs at least 40 GB of disk to buid some ports.

the CI will fail. will try to fix it when more disk space is available
on GitHub Actions runners.